requirements.txt
crewai>=0.28.0
langchain>=0.1.0
langgraph>=0.0.15
groq>=0.4.0
pydantic>=2.4.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.5.0
fastapi>=0.104.0
uvicorn>=0.23.2
langchain_groq